+++ src/ScenarioProcessor.ts
@@ -18,31 +18,31 @@ export class ScenarioProcessor<P extends Feature | Rule> extends PartialListProc
         this.tagProcessor = new TagProcessor<Scenario>(preCompiler);
     }
 
-    public preFilter(e: Scenario, p: P, i: number): boolean {
+    public async preFilter(e: Scenario, p: P, i: number): Promise<boolean> {
         /* istanbul ignore next */
         debug(
             "preFilter(hasPreScenario: %s, e: %s, p: %s, i: %d)",
             !!this.preCompiler.preScenario, e?.constructor.name, p?.constructor.name, i
         );
-        return !this.preCompiler.preScenario || this.preCompiler.preScenario(e, p, i);
+        return !this.preCompiler.preScenario || await this.preCompiler.preScenario(e, p, i);
     }
-    public postFilter(e: Scenario, p: P, i: number): boolean {
+    public async postFilter(e: Scenario, p: P, i: number): Promise<boolean> {
         /* istanbul ignore next */
         debug(
-            "postFilter(hasPostScenario: %s, e: %s, p: %s, i: %d)", 
+            "postFilter(hasPostScenario: %s, e: %s, p: %s, i: %d)",
             !!this.preCompiler.postScenario, e?.constructor.name, p?.constructor.name, i
         );
-        return !this.preCompiler.postScenario || this.preCompiler.postScenario(e, p, i);
+        return !this.preCompiler.postScenario || await this.preCompiler.postScenario(e, p, i);
     }
-    public process(e: Scenario, p: P, i: number): MultiControlType<Scenario> {
+    public async process(e: Scenario, p: P, i: number): Promise<MultiControlType<Scenario>> {
         /* istanbul ignore next */
         debug(
-            "process(hasOnScenario: %s, e: %s, p: %s, i: %d)", 
+            "process(hasOnScenario: %s, e: %s, p: %s, i: %d)",
             !!this.preCompiler.onScenario, e?.constructor.name, p?.constructor.name, i
         );
         let scenario: MultiControlType<Scenario> = e;
         if (this.preCompiler.onScenario) {
-            const result = this.preCompiler.onScenario(e, p, i);
+            const result = await this.preCompiler.onScenario(e, p, i);
             if (typeof result !== "undefined") {
                 scenario = result;
             }
@@ -50,19 +50,19 @@ export class ScenarioProcessor<P extends Feature | Rule> extends PartialListProc
         if (scenario) {
             if (Array.isArray(scenario)) {
                 debug("...Array: %d", scenario.length);
-                scenario = scenario.map(this.postProcess.bind(this));
+                scenario = await ScenarioProcessor.map(scenario, this.postProcess.bind(this));
             } else {
                 debug("...replace");
-                scenario = this.postProcess(scenario);
+                scenario = await this.postProcess(scenario);
             }
         }
         return scenario;
     }
 
-    private postProcess(e: Scenario): Scenario {
+    private async postProcess(e: Scenario): Promise<Scenario> {
         debug("postProcess(tags: %s, steps: %s)", Array.isArray(e.tags), Array.isArray(e.steps));
-        e.tags = this.tagProcessor.execute(e.tags, e);
-        e.steps = this.stepProcessor.execute(e.steps, e);
+        e.tags = await this.tagProcessor.execute(e.tags, e);
+        e.steps = await this.stepProcessor.execute(e.steps, e);
         return e;
     }
 }