+++ src/StepProcessor.ts
@@ -18,31 +18,31 @@ export class StepProcessor<P extends Background | Scenario | ScenarioOutline> ex
         this.dataTableProcessor = new DataTableProcessor(preCompiler);
     }
 
-    protected preFilter(e: Step, p: P, i: number): boolean {
+    protected async preFilter(e: Step, p: P, i: number): Promise<boolean> {
         /* istanbul ignore next */
         debug(
             "preFilter(hasPreStep: %s, e: %s, p: %s, i: %d)",
             !!this.preCompiler.preStep, e?.constructor.name, p?.constructor.name, i
         );
-        return !this.preCompiler.preStep || this.preCompiler.preStep(e, p, i);
+        return !this.preCompiler.preStep || await this.preCompiler.preStep(e, p, i);
     }
-    protected postFilter(e: Step, p: P, i: number): boolean {
+    protected async postFilter(e: Step, p: P, i: number): Promise<boolean> {
         /* istanbul ignore next */
         debug(
             "postFilter(hasPostStep: %s, e: %s, p: %s, i: %d)",
             !!this.preCompiler.postStep, e?.constructor.name, p?.constructor.name, i
         );
-        return !this.preCompiler.postStep || this.preCompiler.postStep(e, p, i);
+        return !this.preCompiler.postStep || await this.preCompiler.postStep(e, p, i);
     }
-    protected process(e: Step, p: P, i: number): MultiControlType<Step> {
+    protected async process(e: Step, p: P, i: number): Promise<MultiControlType<Step>> {
         /* istanbul ignore next */
         debug(
-            "process(hasOnStep: %s, e: %s, p: %s, i: %d)", 
+            "process(hasOnStep: %s, e: %s, p: %s, i: %d)",
             !!this.preCompiler.onStep, e?.constructor.name, p?.constructor.name, i
         );
         let step: MultiControlType<Step> = e;
         if (this.preCompiler.onStep) {
-            const result = this.preCompiler.onStep(e, p, i);
+            const result = await this.preCompiler.onStep(e, p, i);
             if (typeof result !== "undefined") {
                 step = result;
             }
@@ -50,24 +50,24 @@ export class StepProcessor<P extends Background | Scenario | ScenarioOutline> ex
         if (step) {
             if (Array.isArray(step)) {
                 debug("...Array: %d", step.length);
-                step = step.map(this.postProcess.bind(this));
+                step = await StepProcessor.map(step, this.postProcess.bind(this));
             } else {
                 debug("...replace");
-                step = this.postProcess(step);
+                step = await this.postProcess(step);
             }
         }
         return step;
     }
 
-    private postProcess(e: Step): Step {
+    private async postProcess(e: Step): Promise<Step> {
         debug("postProcess(docString: %s, dataTable: %s)", !!e.docString, !!e.dataTable);
         if (e.docString) {
             debug("...docString")
-            e.docString = this.docStringProcessor.execute(e.docString, e);
+            e.docString = await this.docStringProcessor.execute(e.docString, e);
         }
         if (e.dataTable) {
             debug("...dataTable")
-            e.dataTable = this.dataTableProcessor.execute(e.dataTable, e);
+            e.dataTable = await this.dataTableProcessor.execute(e.dataTable, e);
         }
         return e;
     }