+++ src/bin/precompiler.ts
@@ -1,11 +1,11 @@
 #!/usr/bin/env node
 import { run } from "../cli";
-import { getDebugger } from "../debug";
-const debug = getDebugger("cli");
 
-run();
-
-process.on("unhandledRejection", error => {
-    debug("unhandledRejection:\n%s", error);
-    process.exit(1);
-});
+(async () => {
+    try {
+        await run();
+    } catch (e) {
+        console.error(e);
+        process.exit(1);
+    }
+})();