+++ tests/DocumentProcessor.test.ts
@@ -12,15 +12,15 @@ describe("DocumentProcessor", () => {
         document2 = new Document("2", feature);
     });
 
-    test("should work if no pre/post-filter or event handler is set", () => {
+    test("should work if no pre/post-filter or event handler is set", async () => {
         const documentProcessor = new DocumentProcessor();
-        const result1 = documentProcessor.execute(document1);
-        const result2 = documentProcessor.execute(document2);
+        const result1 = await documentProcessor.execute(document1);
+        const result2 = await documentProcessor.execute(document2);
         expect(result1).toEqual([]);
         expect(pruneID(result2)).toEqual([pruneID(document2)]);
     });
 
-    test("should handle when processing returns null", () => {
+    test("should handle when processing returns null", async () => {
         const postFeature = jest.fn()
         const documentProcessor = new DocumentProcessor({
             onFeature() {
@@ -28,34 +28,34 @@ describe("DocumentProcessor", () => {
             },
             postFeature,
         });
-        const result = documentProcessor.execute(document2)
+        const result = await documentProcessor.execute(document2)
         expect(result).toEqual([])
     });
 
-    test("should handle when onDocument returns null", () => {
+    test("should handle when onDocument returns null", async () => {
         const documentProcessor = new DocumentProcessor({
             onDocument() {
                 return null;
             }
         });
-        const result = documentProcessor.execute(document2);
+        const result = await documentProcessor.execute(document2);
         expect(result).toEqual([]);
     });
 
-    test("should handle when onDocument modifies document", () => {
+    test("should handle when onDocument modifies document", async () => {
         const documentProcessor = new DocumentProcessor({
             onDocument(d: Document) {
                 d.targetFolder = "TARGET";
             }
         });
-        const result = documentProcessor.execute(document2);
+        const result = await documentProcessor.execute(document2);
         expect(result[0].targetFolder).toEqual("TARGET");
     });
 
-    test("should handle null document", () => {
+    test("should handle null document", async () => {
         const documentProcessor = new DocumentProcessor();
 
-        const result = documentProcessor.execute(null);
+        const result = await documentProcessor.execute(null);
 
         expect(result).toEqual([]);
     });